Fixing initializer when using older versions of ember-cli-babel.
